For LaTeX output, use explicit {} with \_
This improves the rendered output for certain combinations of letters and
diacriticals, such as lowercase alpha + smooth + acute.  I think what
happens without the braces is that XeLaTeX still parses the macro as
applying only to the alpha and then attempts to compose the combining
diacriticals with a box, rather than noticing it can use a precomposed
glyph.